added Gateway and resolved all the issues

+++ BookMyshow-SpringBoot-Microservices/Theatre/src/main/java/com/BookMyShow/Theatre/shows/ShowsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ShowsRepository extends JpaRepository<Shows, Long> {
+public interface ShowsRepository extends JpaRepository<Shows, Integer> {
 //    @Query("SELECT s FROM shows s JOIN theatre t ON s.theatre.theatre_id = s.id WHERE t.name = :theatreName")
     List<Shows> findByTheatreId(int theatreId);
 }
